// src/app/_components/types/mangadex.ts
export type Relationship = {
  /** UUID formatted string */
  id: string;
  type: string;
  /** Only present if you are on a Manga entity and a Manga relationship */
  related:
    | "monochrome"
    | "main_story"
    | "adapted_from"
    | "based_on"
    | "prequel"
    | "side_story"
    | "doujinshi"
    | "same_franchise"
    | "shared_universe"
    | "sequel"
    | "spin_off"
    | "alternate_story"
    | "alternate_version"
    | "preserialization"
    | "colored"
    | "serialization";
  /** If Reference Expansion is applied, contains objects attributes */
  attributes: any | null;
};

export type Chapter = {
  /** UUID formatted string */
  id: string;
  type: "chapter";
  attributes: ChapterAttributes;
  relationships: Relationship[];
};

export type Manga = {
  /** UUID formatted string */
  id: string;
  type: "manga";
  attributes: MangaAttributes;
  relationships: Relationship[];
};

export type ErrorResponse = {
  /** Default: "error" */
  result: string;
  errors: Error[];
};

export type Error = {
  id: string;
  status: number;
  title: string;
  detail: string;
};

export type ChapterAttributes = {
  /** ```console
   * Maximum length: 255
   * ``` */
  title: string | null;
  volume: string | null;
  /** ```console
   * Maximum length: 8
   * ``` */
  chapter: string | null;
  pages: number;
  /** Pattern: `^[a-z]{2}(-[a-z]{2})?$` */
  translatedLanguage: string;
  /** UUID formatted string */
  uploader: string;
  /** ```console
   * Maximum length: 512
   * Pattern: ^https?://
   * ``` */
  externalUrl: string | null;
  /** ```console
   * Minimum: 1
   * ``` */
  version: number;
  createdAt: string;
  updatedAt: string;
  publishAt: string;
  readableAt: string;
};

export type MangaAttributes = {
  title: LocalizedString;
  altTitles: LocalizedString[];
  description: LocalizedString;
  isLocked: boolean;
  links: Links | null;
  originalLanguage: string;
  lastVolume: string | null;
  lastChapter: string | null;
  publicationDemographic: "shounen" | "shoujo" | "josei" | "seinen" | null;
  status: "completed" | "ongoing" | "cancelled" | "hiatus";
  year: number | null;
  contentRating: "safe" | "suggestive" | "erotica" | "pornographic";
  chapterNumbersResetOnNewVolume: boolean;
  availableTranslatedLanguages: any[];
  /** UUID formatted string */
  latestUploadedChapter: string;
  tags: Tag[];
  state: "draft" | "submitted" | "published" | "rejected";
  /** ```console
   * Minimum: 1
   * ``` */
  version: number;
  createdAt: string;
  updatedAt: string;
};
// ...
export type ScanlationGroup = {
  /** UUID formatted string */
  id: string;
  type: "scanlation_group";
  attributes: ScanlationGroupAttributes;
  relationships: Relationship[];
};

export type ScanlationGroupAttributes = {
  name: string;
  altNames: LocalizedString[];
  website: string | null;
  ircServer: string | null;
  ircChannel: string | null;
  discord: string | null;
  contactEmail: string | null;
  description: string | null;
  /** Pattern: `^https?://` */
  twitter: string | null;
  /** ```console
   * Maximum length: 128
   * Pattern: ^https:\/\/www\.mangaupdates\.com\/(group|publisher)(s\.html\?id=\d+|\/[\w-]+\/?([\w-]+)?(\/)?)$
   */
  mangaUpdates: string | null;
  /** Pattern: `^[a-z]{2}(-[a-z]{2})?$` */
  focusedLanguages: string[] | null;
  locked: boolean;
  official: boolean;
  inactive: boolean;
  /**
   * Should respect ISO 8601 duration specifications: https://en.wikipedia.org/wiki/ISO_8601#Durations
   *
   * Pattern: `^(P([1-9]|[1-9][0-9])D)?(P?([1-9])W)?(P?T(([1-9]|1[0-9]|2[0-4])H)?(([1-9]|[1-5][0-9]|60)M)?(([1-9]|[1-5][0-9]|60)S)?)?$`
   */
  publishDelay: string;
  /** ```console
   * Minimum: 1
   * ``` */
  version: number;
  createdAt: string;
  updatedAt: string;
};
// ...
export type ExtendChapter = Chapter & {
  manga?: Partial<Manga> & Pick<Manga, "id" | "type">;
  scanlation_group?: Partial<ScanlationGroup> &
    Pick<ScanlationGroup, "id" | "type">;
};